Starting to fetch heroes
